TruthLens-AI/requirements.txt
# TruthLens Intelligence - Fake News Detection System
# Python 3.11+

# Machine Learning
scikit-learn>=1.3.0
nltk>=3.8.1
pandas>=2.0.0
numpy>=1.24.0

# Backend
flask>=3.0.0
PyJWT>=2.0.0
pymongo>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Note: Pickle is part of Python standard library
